+++ client/src/pages/DetallesPage.tsx
@@ -2,36 +2,19 @@ import { Card, Select, SelectItem, Table, TableBody, TableCell, TableHead, Table
 import { RiNumbersLine, RiDatabase2Fill } from '@remixicon/react'
 import { type Sucursales } from '../types/sucursal'
 import { Input, Label } from '../components/iu'
+import { useFilter } from '../hooks/useFilters'
 import { useEffect, useState } from 'react'
 import axios from 'axios'
 
-interface FilterItems {
-  searchItems: string
-  setSearchItems: React.Dispatch<React.SetStateAction<string>>
-  filteredItems: Sucursales
-}
-
-function useFilter (items: Sucursales): FilterItems {
-  const [searchItems, setSearchItems] = useState('')
-
-  const filteredItems =
-  items.filter(({ PDV_NOMBRE, PDV_SUCURSAL }) =>
-    PDV_NOMBRE.toLowerCase().includes(searchItems.toLowerCase()) ||
-    PDV_SUCURSAL.toString().toLowerCase().includes(searchItems.toLowerCase())
-  )
-
-  return { searchItems, setSearchItems, filteredItems }
-}
-
 const DetallesPage = (): JSX.Element => {
   const [data, setData] = useState<Sucursales>([])
+
+  const { filteredItems, searchItems, setSearchItems } = useFilter(data)
 
   useEffect(() => {
     void axios.get('http://localhost:3000/api/sucursales')
       .then((response) => { setData(response.data as Sucursales) })
   }, [])
-
-  const { filteredItems, searchItems, setSearchItems } = useFilter(data)
 
   return (
     <>
